backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
marshmallow
python-dotenv
psycopg2-binary
PyJWT
ccxt
logging
requests
flask-cors
pandas